Make the Gradle plugin work without the Kotlin Plugin (#3496)
* add a Java-Only project

* move the Kotlin plugin code to KotlinPluginFacade

* update apiDump

* isolate Android projects too

* optimize imports

* update apiDump

* add missing space in error message
