+++ apollo-gradle-plugin/src/main/kotlin/com/apollographql/apollo3/gradle/api/Service.kt
@@ -349,7 +349,7 @@ interface Service {
    *
    * Default: listOf(".*")
    */
-  @Deprecated("Kotlin sealed classes are more flexible than Kotlin enums to represent GraphQL enums because they can expose the" +
+  @Deprecated("Kotlin sealed classes are more flexible than Kotlin enums to represent GraphQL enums because they can expose the " +
       "rawValue of the unknown enums.")
   val sealedClassesForEnumsMatching: ListProperty<String>
 
@@ -469,8 +469,12 @@ interface Service {
     /**
      * Connects the generated sources to the given Android variant. This will
      * look up the most specific source set used by this variant. For an example, "demoDebug"
+     *
+     * @param variant: the [BaseVariant] to connect to. It is of type [Any] because [DirectoryConnection]
+     * can be used in non-Android projects, and we don't want the class to fail during loading because
+     * of a missing symbol in that case
      */
-    fun connectToAndroidVariant(variant: BaseVariant)
+    fun connectToAndroidVariant(variant: Any)
 
     /**
      * The directory where the generated models will be written.
