+++ apollo-gradle-plugin/src/main/kotlin/com/apollographql/apollo3/gradle/internal/DefaultDirectoryConnection.kt
@@ -1,28 +1,13 @@
 package com.apollographql.apollo3.gradle.internal
 
-import com.android.build.gradle.AppExtension
-import com.android.build.gradle.FeatureExtension
-import com.android.build.gradle.LibraryExtension
-import com.android.build.gradle.TestedExtension
-import com.android.build.gradle.api.AndroidSourceSet
-import com.android.build.gradle.api.BaseVariant
-import com.android.build.gradle.api.TestVariant
-import com.android.build.gradle.api.UnitTestVariant
-import com.apollographql.apollo3.compiler.capitalizeFirstLetter
 import com.apollographql.apollo3.gradle.api.Service
-import com.apollographql.apollo3.gradle.api.androidExtensionOrThrow
 import com.apollographql.apollo3.gradle.api.javaConventionOrThrow
 import com.apollographql.apollo3.gradle.api.kotlinProjectExtensionOrThrow
 import org.gradle.api.Project
 import org.gradle.api.Task
 import org.gradle.api.file.Directory
-import org.gradle.api.file.SourceDirectorySet
-import org.gradle.api.internal.HasConvention
 import org.gradle.api.provider.Provider
 import org.gradle.api.tasks.TaskProvider
-import org.jetbrains.kotlin.gradle.plugin.KOTLIN_DSL_NAME
-import org.jetbrains.kotlin.gradle.plugin.KOTLIN_JS_DSL_NAME
-import org.jetbrains.kotlin.gradle.plugin.KotlinSourceSet
 
 internal class DefaultDirectoryConnection(
     private val project: Project,
@@ -41,82 +26,11 @@ internal class DefaultDirectoryConnection(
         .srcDir(outputDir)
   }
 
-  private fun Any.getConvention(name: String): Any? =
-      (this as HasConvention).convention.plugins[name]
-
-  // Copied from kotlin plugin
-  private val AndroidSourceSet.kotlinSourceSet: SourceDirectorySet?
-    get() {
-      val convention = (getConvention(KOTLIN_DSL_NAME) ?: getConvention(KOTLIN_JS_DSL_NAME)) ?: return null
-      val kotlinSourceSetIface =
-          convention.javaClass.interfaces.find { it.name == KotlinSourceSet::class.qualifiedName }
-      val getKotlin = kotlinSourceSetIface?.methods?.find { it.name == "getKotlin" } ?: return null
-      return getKotlin(convention) as? SourceDirectorySet
-    }
-
-  override fun connectToAndroidVariant(variant: BaseVariant) {
-    /**
-     * Heuristic to get the variant-specific sourceSet from the variant name
-     * demoDebugAndroidTest -> androidTestDemoDebug
-     * demoDebugUnitTest -> testDemoDebug
-     * demoDebug -> demoDebug
-     */
-    val sourceSetName = when {
-      variant is TestVariant && variant.name.endsWith("AndroidTest") -> {
-        "androidTest${variant.name.removeSuffix("AndroidTest").capitalizeFirstLetter()}"
-      }
-      variant is UnitTestVariant && variant.name.endsWith("UnitTest") -> {
-        "test${variant.name.removeSuffix("UnitTest").capitalizeFirstLetter()}"
-      }
-      else -> variant.name
-    }
-    connectToAndroidSourceSet(sourceSetName)
+  override fun connectToAndroidVariant(variant: Any) {
+    connectToAndroidVariant(project, variant, outputDir)
   }
 
   override fun connectToAndroidSourceSet(name: String) {
-    val container = project.container(BaseVariant::class.java)
-
-    val extension = project.androidExtensionOrThrow
-    when (extension) {
-      is LibraryExtension -> {
-        extension.libraryVariants.all { variant ->
-          container.add(variant)
-        }
-      }
-      is AppExtension -> {
-        extension.applicationVariants.all { variant ->
-          container.add(variant)
-        }
-      }
-      is FeatureExtension -> {
-        extension.featureVariants.all { variant ->
-          container.add(variant)
-        }
-      }
-      else -> error("Unsupported extension: $extension")
-    }
-
-    if (extension is TestedExtension) {
-      extension.testVariants.all { variant ->
-        container.add(variant)
-      }
-      extension.unitTestVariants.all { variant ->
-        container.add(variant)
-      }
-    }
-
-    container.all {
-      if (it.sourceSets.any { it.name == name }) {
-        // This is required for AS to see the sources
-        // See https://github.com/apollographql/apollo-android/issues/3351
-        it.addJavaSourceFoldersToModel(outputDir.get().asFile)
-      }
-    }
-
-    project.androidExtensionOrThrow
-        .sourceSets
-        .getByName(name)
-        .kotlinSourceSet!!
-        .srcDir(outputDir)
+    connectToAndroidSourceSet(project, name, outputDir)
   }
 }
