+++ apollo-gradle-plugin/testProjects/java-only/src/main/graphql/operation.graphql
@@ -1,0 +1,3 @@
+query GetRandom {
+    random
+}
