+++ apollo-gradle-plugin/testProjects/java-only/src/main/graphql/schema.graphqls
@@ -1,0 +1,3 @@
+type Query {
+    random: String
+}
